+++ connector/connect/client/jvm-internal/src/test/scala/org/apache/spark/sql/connect/client/util/ConnectFunSuite.scala
@@ -18,12 +18,16 @@ package org.apache.spark.sql.connect.client.util
 
 import java.nio.file.Path
 
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
 import org.scalatest.funsuite.AnyFunSuite // scalastyle:ignore funsuite
 
 /**
  * The basic testsuite the client tests should extend from.
  */
-trait ConnectFunSuite extends AnyFunSuite { // scalastyle:ignore funsuite
+trait ConnectFunSuite
+  extends AnyFunSuite // scalastyle:ignore funsuite
+  with BeforeAndAfterAll
+  with BeforeAndAfterEach {
 
   // Borrowed from SparkFunSuite
   protected def getWorkspaceFilePath(first: String, more: String*): Path = {
