+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/connect/client/SparkConnectClientSuite.scala
@@ -19,19 +19,14 @@ package org.apache.spark.sql.connect.client
 import java.util.UUID
 import java.util.concurrent.TimeUnit
 
-import scala.collection.JavaConverters._
-import scala.collection.mutable
-
 import io.grpc.{CallOptions, Channel, ClientCall, ClientInterceptor, MethodDescriptor, Server, Status, StatusRuntimeException}
 import io.grpc.netty.NettyServerBuilder
-import io.grpc.stub.StreamObserver
 import org.scalatest.BeforeAndAfterEach
 
 import org.apache.spark.SparkException
 import org.apache.spark.connect.proto
-import org.apache.spark.connect.proto.{AddArtifactsRequest, AddArtifactsResponse, AnalyzePlanRequest, AnalyzePlanResponse, ArtifactStatusesRequest, ArtifactStatusesResponse, ExecutePlanRequest, ExecutePlanResponse, SparkConnectServiceGrpc}
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.connect.client.util.ConnectFunSuite
+import org.apache.spark.sql.connect.client.util.{ConnectFunSuite, DummySparkConnectService}
 import org.apache.spark.sql.connect.common.config.ConnectCommon
 
 class SparkConnectClientSuite extends ConnectFunSuite with BeforeAndAfterEach {
@@ -77,7 +72,7 @@ class SparkConnectClientSuite extends ConnectFunSuite with BeforeAndAfterEach {
       clientBuilder: Int => SparkConnectClient): Unit = {
     startDummyServer(serverPort)
     client = clientBuilder(server.getPort)
-    val request = AnalyzePlanRequest
+    val request = proto.AnalyzePlanRequest
       .newBuilder()
       .setSessionId("abc123")
       .build()
@@ -104,7 +99,7 @@ class SparkConnectClientSuite extends ConnectFunSuite with BeforeAndAfterEach {
       .retryPolicy(GrpcRetryHandler.RetryPolicy(maxRetries = 0))
       .build()
 
-    val request = AnalyzePlanRequest.newBuilder().setSessionId("abc123").build()
+    val request = proto.AnalyzePlanRequest.newBuilder().setSessionId("abc123").build()
 
     // Failed the ssl handshake as the dummy server does not have any server credentials installed.
     assertThrows[SparkException] {
@@ -273,151 +268,5 @@ class SparkConnectClientSuite extends ConnectFunSuite with BeforeAndAfterEach {
       retryHandler.retry { dummyFn.fn() }
     }
     assert(dummyFn.counter == 2)
-  }
-}
-
-class DummySparkConnectService() extends SparkConnectServiceGrpc.SparkConnectServiceImplBase {
-
-  private var inputPlan: proto.Plan = _
-  private val inputArtifactRequests: mutable.ListBuffer[AddArtifactsRequest] =
-    mutable.ListBuffer.empty
-
-  private[sql] def getAndClearLatestInputPlan(): proto.Plan = {
-    val plan = inputPlan
-    inputPlan = null
-    plan
-  }
-
-  private[sql] def getAndClearLatestAddArtifactRequests(): Seq[AddArtifactsRequest] = {
-    val requests = inputArtifactRequests.toSeq
-    inputArtifactRequests.clear()
-    requests
-  }
-
-  override def executePlan(
-      request: ExecutePlanRequest,
-      responseObserver: StreamObserver[ExecutePlanResponse]): Unit = {
-    // Reply with a dummy response using the same client ID
-    val requestSessionId = request.getSessionId
-    val operationId = if (request.hasOperationId) {
-      request.getOperationId
-    } else {
-      UUID.randomUUID().toString
-    }
-    inputPlan = request.getPlan
-    val response = ExecutePlanResponse
-      .newBuilder()
-      .setSessionId(requestSessionId)
-      .setOperationId(operationId)
-      .build()
-    responseObserver.onNext(response)
-    // Reattachable execute must end with ResultComplete
-    if (request.getRequestOptionsList.asScala.exists { option =>
-        option.hasReattachOptions && option.getReattachOptions.getReattachable == true
-      }) {
-      val resultComplete = ExecutePlanResponse
-        .newBuilder()
-        .setSessionId(requestSessionId)
-        .setOperationId(operationId)
-        .setResultComplete(proto.ExecutePlanResponse.ResultComplete.newBuilder().build())
-        .build()
-      responseObserver.onNext(resultComplete)
-    }
-    responseObserver.onCompleted()
-  }
-
-  override def analyzePlan(
-      request: AnalyzePlanRequest,
-      responseObserver: StreamObserver[AnalyzePlanResponse]): Unit = {
-    // Reply with a dummy response using the same client ID
-    val requestSessionId = request.getSessionId
-    request.getAnalyzeCase match {
-      case proto.AnalyzePlanRequest.AnalyzeCase.SCHEMA =>
-        inputPlan = request.getSchema.getPlan
-      case proto.AnalyzePlanRequest.AnalyzeCase.EXPLAIN =>
-        inputPlan = request.getExplain.getPlan
-      case proto.AnalyzePlanRequest.AnalyzeCase.TREE_STRING =>
-        inputPlan = request.getTreeString.getPlan
-      case proto.AnalyzePlanRequest.AnalyzeCase.IS_LOCAL =>
-        inputPlan = request.getIsLocal.getPlan
-      case proto.AnalyzePlanRequest.AnalyzeCase.IS_STREAMING =>
-        inputPlan = request.getIsStreaming.getPlan
-      case proto.AnalyzePlanRequest.AnalyzeCase.INPUT_FILES =>
-        inputPlan = request.getInputFiles.getPlan
-      case _ => inputPlan = null
-    }
-    val response = AnalyzePlanResponse
-      .newBuilder()
-      .setSessionId(requestSessionId)
-      .build()
-    responseObserver.onNext(response)
-    responseObserver.onCompleted()
-  }
-
-  override def addArtifacts(responseObserver: StreamObserver[AddArtifactsResponse])
-      : StreamObserver[AddArtifactsRequest] = new StreamObserver[AddArtifactsRequest] {
-    override def onNext(v: AddArtifactsRequest): Unit = inputArtifactRequests.append(v)
-
-    override def onError(throwable: Throwable): Unit = responseObserver.onError(throwable)
-
-    override def onCompleted(): Unit = {
-      responseObserver.onNext(proto.AddArtifactsResponse.newBuilder().build())
-      responseObserver.onCompleted()
-    }
-  }
-
-  override def artifactStatus(
-      request: ArtifactStatusesRequest,
-      responseObserver: StreamObserver[ArtifactStatusesResponse]): Unit = {
-    val builder = proto.ArtifactStatusesResponse.newBuilder()
-    request.getNamesList().iterator().asScala.foreach { name =>
-      val status = proto.ArtifactStatusesResponse.ArtifactStatus.newBuilder()
-      val exists = if (name.startsWith("cache/")) {
-        inputArtifactRequests.exists { artifactReq =>
-          if (artifactReq.hasBatch) {
-            val batch = artifactReq.getBatch
-            batch.getArtifactsList.asScala.exists { singleArtifact =>
-              singleArtifact.getName == name
-            }
-          } else false
-        }
-      } else false
-      builder.putStatuses(name, status.setExists(exists).build())
-    }
-    responseObserver.onNext(builder.build())
-    responseObserver.onCompleted()
-  }
-
-  override def interrupt(
-      request: proto.InterruptRequest,
-      responseObserver: StreamObserver[proto.InterruptResponse]): Unit = {
-    val response = proto.InterruptResponse.newBuilder().setSessionId(request.getSessionId).build()
-    responseObserver.onNext(response)
-    responseObserver.onCompleted()
-  }
-
-  override def reattachExecute(
-      request: proto.ReattachExecuteRequest,
-      responseObserver: StreamObserver[proto.ExecutePlanResponse]): Unit = {
-    // Reply with a dummy response using the same client ID
-    val requestSessionId = request.getSessionId
-    val response = ExecutePlanResponse
-      .newBuilder()
-      .setSessionId(requestSessionId)
-      .build()
-    responseObserver.onNext(response)
-    responseObserver.onCompleted()
-  }
-
-  override def releaseExecute(
-      request: proto.ReleaseExecuteRequest,
-      responseObserver: StreamObserver[proto.ReleaseExecuteResponse]): Unit = {
-    val response = proto.ReleaseExecuteResponse
-      .newBuilder()
-      .setSessionId(request.getSessionId)
-      .setOperationId(request.getOperationId)
-      .build()
-    responseObserver.onNext(response)
-    responseObserver.onCompleted()
   }
 }
